Start moving about llik functions
