+++ src/Makevars.in
@@ -13,7 +13,7 @@ PKG_CXXFLAGS = -Id -I../inst/include -DBOOST_DISABLE_ASSERTS -DBOOST_NO_CXX11_ST
 SHLIB_LDFLAGS = $(SHLIB_CXXLDFLAGS) 
 SHLIB_LD = $(SHLIB_CXXLD)
 SOURCES_C = init.c rprintf.c merge3.c lbfgsR.c utilc.c
-SOURCES_CPP = saem.cpp inner.cpp expandNlme.cpp npde.cpp shrink.cpp ires.cpp res.cpp cwres.cpp censResid.cpp cholse.cpp neldermead.cpp resid.cpp slice.cpp RcppExportMod.cpp ode_cmt1.cpp RcppExports.cpp
+SOURCES_CPP = llik.cpp saem.cpp inner.cpp expandNlme.cpp npde.cpp shrink.cpp ires.cpp res.cpp cwres.cpp censResid.cpp cholse.cpp neldermead.cpp resid.cpp slice.cpp RcppExportMod.cpp ode_cmt1.cpp RcppExports.cpp
 OBJECTS = $(SOURCES_CPP:.cpp=.o) $(SOURCES_C:.c=.o) 
 
 all: $(SHLIB)
